Update Push register API
