+++ Source/BMSPushClient.swift
@@ -761,6 +761,60 @@ import BMSCore
             }
         }
         
+        public func setupPush ()  {
+            
+            if #available(iOS 10.0, *) {
+                UNUserNotificationCenter.current().requestAuthorization(options: [.alert, .badge, .sound], completionHandler: { (granted, error) in
+                    if(granted) {
+                        UIApplication.shared.registerForRemoteNotifications()
+                    } else {
+                        print("Error while registering with APNS server :  \(error?.localizedDescription)")
+                    }
+                })
+            } else {
+                // Fallback on earlier versions
+                
+                let options :BMSPushClientOptions  = BMSPushClient.sharedInstance.notificationOptions!
+                let category : [BMSPushNotificationActionCategory] = options.category
+                
+                let categoryFirst : BMSPushNotificationActionCategory = category.first!
+                
+                let pushNotificationAction : [BMSPushNotificationAction] = categoryFirst.actions
+                let pushCategoryIdentifier : String = categoryFirst.identifier
+                
+                let firstActionButton : BMSPushNotificationAction = pushNotificationAction.first!
+                let secondActionButton : BMSPushNotificationAction = pushNotificationAction[1]
+                
+                let replyActionButtonOne : UIMutableUserNotificationAction = UIMutableUserNotificationAction()
+                replyActionButtonOne.identifier = firstActionButton.identifier
+                replyActionButtonOne.title = firstActionButton.title
+                replyActionButtonOne.activationMode = firstActionButton.activationMode
+                replyActionButtonOne.isAuthenticationRequired = firstActionButton.authenticationRequired!
+                
+                let replyActionButtonTwo : UIMutableUserNotificationAction = UIMutableUserNotificationAction()
+                replyActionButtonTwo.identifier = secondActionButton.identifier
+                replyActionButtonTwo.title = secondActionButton.title
+                replyActionButtonTwo.activationMode = secondActionButton.activationMode
+                replyActionButtonTwo.isAuthenticationRequired = secondActionButton.authenticationRequired!
+                
+                let responseCategory : UIMutableUserNotificationCategory = UIMutableUserNotificationCategory()
+                responseCategory.identifier = pushCategoryIdentifier
+                
+                let replyActions: [UIUserNotificationAction] = [replyActionButtonOne, replyActionButtonTwo]
+                
+                responseCategory.setActions(replyActions, for:UIUserNotificationActionContext.default)
+                responseCategory.setActions(replyActions, for:UIUserNotificationActionContext.minimal)
+                
+                let categories = NSSet(object: responseCategory)
+                
+                let settings = UIUserNotificationSettings(types: [.alert, .badge, .sound], categories: categories as! Set<UIUserNotificationCategory>)
+                
+                UIApplication.shared.registerUserNotificationSettings(settings)
+                UIApplication.shared.registerForRemoteNotifications()
+                
+            }
+        }
+        
         internal func validateString(object:String) -> Bool{
             if (object.isEmpty || object == "") {
                 return false;
@@ -1476,11 +1530,42 @@ import BMSCore
                 }
             })
         }
-        
+    
+    public func setupPush() {
+    
+    let replyActionButtonOne : UIMutableUserNotificationAction = UIMutableUserNotificationAction()
+    replyActionButtonOne.identifier = "FIRST_BUTTON"
+    replyActionButtonOne.title = "First"
+    replyActionButtonOne.activationMode = UIUserNotificationActivationMode.background
+    replyActionButtonOne.isAuthenticationRequired = false
+    
+    let replyActionButtonTwo : UIMutableUserNotificationAction = UIMutableUserNotificationAction()
+    replyActionButtonTwo.identifier = "SECOND_BUTTON"
+    replyActionButtonTwo.title = "Second"
+    replyActionButtonTwo.activationMode = UIUserNotificationActivationMode.background
+    replyActionButtonTwo.isAuthenticationRequired = false
+    
+    let responseCategory : UIMutableUserNotificationCategory = UIMutableUserNotificationCategory()
+    responseCategory.identifier = "categoryOne"
+    
+    //let replyActions: NSArray = [replyActionButtonOne, replyActionButtonTwo]
+    let replyActions: [UIUserNotificationAction] = [replyActionButtonOne, replyActionButtonTwo]
+    
+    responseCategory.setActions(replyActions, for:UIUserNotificationActionContext.default)
+    responseCategory.setActions(replyActions, for:UIUserNotificationActionContext.minimal)
+    
+    let categories = NSSet(object: responseCategory)
+    
+    let settings = UIUserNotificationSettings(types: [.alert, .badge, .sound], categories: categories as! Set<UIUserNotificationCategory>)
+    
+    UIApplication.shared.registerUserNotificationSettings(settings)
+    UIApplication.shared.registerForRemoteNotifications()
+    }
+    
         // MARK: Methods (Internal)
-        
+    
         //Begin Logger implementation
-        
+    
         // Setting Log info
         internal func sendAnalyticsData (logType:LogLevel, logStringData:String){
             var devId = String()
